fix with_column_sql bug, update sql_utils
In this function, the sql expression is rewritten to select from 'batch_arrow', so an error is returned if another table name is referenced during the sql_expression. This PR rewrites the expression to change table name references to 'batch_arrow' and also includes some other updates to sql_utils.